Allow user to close filters
