+++ app/src/pages/request-logs/index.tsx
@@ -1,11 +1,9 @@
 import { Text, VStack, HStack, Box, IconButton, Icon, keyframes } from "@chakra-ui/react";
 import { BiRefresh } from "react-icons/bi";
-import { FiFilter } from "react-icons/fi";
 
 import AppShell from "~/components/nav/AppShell";
 import LoggedCallTable from "~/components/requestLogs/LoggedCallsTable";
 import LoggedCallsPaginator from "~/components/requestLogs/LoggedCallsPaginator";
-import ActionButton from "~/components/ActionButton";
 import LogFilters from "~/components/requestLogs/LogFilters";
 import ColumnVisibilityDropdown from "~/components/requestLogs/ColumnVisibilityDropdown";
 import ExportButton from "~/components/requestLogs/ExportButton";
@@ -13,6 +11,7 @@ import AddToDatasetButton from "~/components/requestLogs/AddToDatasetButton";
 import { api } from "~/utils/api";
 import { useLoggedCalls } from "~/utils/hooks";
 import { useFilters } from "~/components/Filters/useFilters";
+import ToggleFiltersButton from "~/components/ToggleFiltersButton";
 
 const spin = keyframes`
   from { transform: rotate(0deg); }
@@ -21,7 +20,6 @@ const spin = keyframes`
 
 export default function LoggedCalls() {
   const filtersShown = useFilters(true).filtersShown;
-  const setFiltersShown = useFilters(true).setFiltersShown;
 
   const utils = api.useContext();
 
@@ -54,13 +52,7 @@ export default function LoggedCalls() {
             <AddToDatasetButton />
             <ExportButton />
             <ColumnVisibilityDropdown />
-            <ActionButton
-              onClick={() => {
-                setFiltersShown(!filtersShown);
-              }}
-              label={filtersShown ? "Hide Filters" : "Show Filters"}
-              icon={FiFilter}
-            />
+            <ToggleFiltersButton defaultShown />
           </HStack>
           {filtersShown && <LogFilters />}
           <LoggedCallTable />
